requirements.txt
alembic==1.7.6
altgraph==0.17.2
anyio==3.5.0
apache-airflow==2.2.4
apache-airflow-providers-databricks==2.2.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.3
apache-airflow-providers-imap==2.2.0
apache-airflow-providers-sqlite==2.1.0
apispec==3.3.2
Appium-Python-Client==2.1.0
argcomplete==2.0.0
asn1crypto==1.4.0
async-generator==1.10
attrs==20.3.0
auto-py-to-exe==2.15.0
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blinker==1.4
boto3==1.15.18
botocore==1.18.18
bottle==0.12.19
bottle-websocket==0.2.9
cachelib==0.6.0
cattrs==1.10.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
connexion==2.12.0
croniter==1.3.4
cryptography==3.4.8
cx-Oracle==8.3.0
cycler==0.11.0
datacompy==0.7.3
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.0
docutils==0.16
dominate==2.6.0
Eel==0.12.4
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
Flask==1.1.4
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-Excel==0.0.7
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.29.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
graphviz==0.19.1
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==2.10
importlib-metadata==4.11.2
importlib-resources==5.4.0
imutils==0.5.4
inflection==0.5.1
iso8601==1.0.2
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpath==0.82
jsonschema==3.2.0
jupyter-core==4.7.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lml==0.1.0
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
msrest==0.6.19
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.0
opencv-python==4.5.5.62
openpyxl==3.0.9
ordered-set==4.0.2
oscrypto==1.2.1
outcome==1.1.0
packaging==21.3
pandas==1.3.2
parso==0.8.2
pathlib==1.0.1
patsy==0.5.2
pdf2image==1.16.0
pefile==2021.9.3
pendulum==2.1.2
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
poppler-utils==0.1.0
prison==0.2.1
prompt-toolkit==3.0.20
psutil==5.9.0
psycopg2==2.9.1
py4j==0.10.9.2
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.8
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
Pygments==2.10.0
pyinstaller==4.9
pyinstaller-hooks-contrib==2021.3
pyjsparser==2.7.1
PyJWT==1.7.1
PyMuPDF==1.19.5
pyodbc==4.0.32
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyspark==3.2.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.27.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==11.2.0
s3transfer==0.3.3
scipy==1.7.1
selenium==4.1.0
semantic-version==2.8.5
sentry-sdk==1.5.7
setproctitle==1.2.2
setuptools-rust==0.12.1
six==1.16.0
sniffio==1.2.0
snowflake==0.0.3
snowflake-connector-python==2.7.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
statsmodels==0.13.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
teradata==15.10.0.21
teradatasql==17.10.0.2
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
texttohtml==1.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.1
tzlocal==3.0
unicodecsv==0.14.1
urllib3==1.25.11
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.13.3
wsproto==1.0.0
WTForms==2.3.3
XlsxWriter==3.0.2
xlwings==0.25.3
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0